+++ model_fetch/lib/src/abstract_filter.dart
@@ -3,6 +3,8 @@ import 'dart:convert';
 abstract class AbstractFilter {
   const AbstractFilter();
 
+  int get pageSize => 20;
+
   Map<String, dynamic> toJson() => const {};
 
   String get hash => jsonEncode(toJson());